Make it actually work, add non-assemble option (-S)

+++ src/brainfuck.rs
@@ -2,59 +2,57 @@ use std::fs;
 use std::process;
 
 // Accepts utf-8 string with only brainfuck characters, ignores everything else
-pub fn compile(path: &str, file: Option<&str>) {
+pub fn compile(path: &str, file: Option<&str>, assemble: bool) {
     let code: String = fs::read_to_string(path).unwrap();
-    
+
     let mut output = String::new();
     output += "global _start\n";
     output += "section .bss\n";
-    output += "memory: resb 65536\n"; // Declare memory space
+    output += "m: resb 65536\n"; // Declare memory space
     output += "section .text\n";
     output += "_start:\n";
-    output += "xor rbx, rbx\n"; //Set initial position to 0
+    output += "mov rbx, 0\n"; // Set mem pointer to memory
 
-    let mut open: usize = 0;
-    let mut close: usize = 0;
+    let mut open: usize = 100; // skip o16, o32, o64 etc
     let mut list: Vec<usize> = Vec::new();
 
     for c in code.chars() {
         match c {
-            '<' => output += "sub bx, 1\n",
-            '>' => output += "add bx, 1\n",
-            '+' => output += "add byte[memory + rbx], 1\n",
-            '-' => output += "sub byte[memory + rbx], 1\n",
+            '<' => output += "dec bx\n",
+            '>' => output += "inc bx\n",
+            '+' => output += "inc byte[m+rbx]\n",
+            '-' => output += "dec byte[m+rbx]\n",
             '[' => {
                 open += 1;
                 list.push(open);
-
-                output += &format!("open_{}:\n", open.to_string());
-                output += "cmp byte[memory + rbx], 0\n";
-                output += &format!("jz close_{}\n", open.to_string());
+                output += &format!("jmp c{}\n", open.to_string());
+                output += &format!("o{}:\n", open.to_string());
             },
             ']' => {
-                close += 1;
                 let Some(jump) = list.pop() else { // If list is empty, we have run out of opening brackets to jmp to
                     println!("Bracket mismatch error. Closing bracket ']' found with no matching opening bracket '['");
                     process::exit(1);
                 };
 
-                output += &format!("jmp open_{}\n", jump.to_string()); // Always go back, because if it's 0 it will jump right back to close
-                output += &format!("close_{}:\n", close.to_string());
+                output += &format!("c{}:\n", jump.to_string());
+                output += "cmp byte[m+rbx], 0\n";
+                output += &format!("jnz o{}\n", jump.to_string());
+
             },
             '.' => {
                 output += "mov rax, 1\n"; // Write syscall
                 output += "mov rdi, 1\n"; // fd = stdout
                 output += "mov rdx, 1\n"; // size = 1
-                output += "mov rsi, memory\n";
-                output += "add rsi, rbx\n";
+                output += "mov rsi, m\n"; // memory pointer
+                output += "add rsi, rbx\n"; // memory offset
                 output += "syscall\n";
             },
             ',' => {
                 output += "mov rax, 0\n"; // Read syscall
                 output += "mov rdi, 0\n"; // fd = stdin
                 output += "mov rdx, 1\n"; // size = 1
-                output += "mov rsi, memory\n";
-                output += "add rsi, rbx\n";
+                output += "mov rsi, m\n"; // memory pointer
+                output += "add rsi, rbx\n"; // memory offset
                 output += "syscall\n";
             },
             _ => (),
@@ -62,31 +60,33 @@ pub fn compile(path: &str, file: Option<&str>) {
     }
 
     // Every bracket must have a match
-    if open != close {
+    if list.len() != 0 {
         println!("Not enough closing brackets ']'. Must have one per opening bracket '['.");
         process::exit(2);
     }
 
     // Exit boilerplate
-    output += "movzx rdi, byte [memory + rbx]\n";
+    output += "mov rdi, 0\n";
     output += "mov rax, 60\n";
     output += "syscall\n";
 
     let name = &path[..path.len()-2];
+    let mut to_be_renamed = String::from(name) + "s";
+    fs::write(&to_be_renamed, output).unwrap();
 
-    let assembly = String::from(name) + "s";
-    let object = String::from(name) + "o";
+    if assemble {
+        let object = String::from(name) + "o";
+        process::Command::new("nasm").arg("-felf64").arg(&to_be_renamed).status().expect("Nasm failed to assemble");
+        fs::remove_file(&to_be_renamed).expect("Failed to remove assembly file");
 
-    fs::write(&assembly, output).unwrap();
+        process::Command::new("ld").arg(&object).status().expect("Ld failed to link");
+        fs::remove_file(&object).expect("Failed to remove object file");
 
-    process::Command::new("nasm").arg("-felf64").arg(&assembly).status().expect("Nasm failed to assemble");
-    fs::remove_file(&assembly).expect("Failed to remove assembly file");
-
-    process::Command::new("ld").arg(&object).status().expect("Ld failed to link");
-    fs::remove_file(&object).expect("Failed to remove object file");
+        to_be_renamed = "a.out".to_string();
+    }
 
     match file {
         None => (),
-        Some(o) => fs::rename(&path, o).expect("Failed to rename output file"),
+        Some(out) => fs::rename(&to_be_renamed, out).expect("Failed to rename output file"),
     }
 }
